+++ custom_components/roomflow/const.py
@@ -1,5 +1,5 @@
 DOMAIN = "roomflow"
-VERSION = "0.0.5"
+VERSION = "0.0.6"
 STORAGE_KEY = "roomflow.rooms"
 STORAGE_VERSION = 1